+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Don't treat `package.json` as a project marker.
 * [#987](https://github.com/bbatsov/projectile/issues/987): projectile-ag ignores ag-ignore-list when projectile-project-vcs is git
 * [#1119](https://github.com/bbatsov/projectile/issues/1119): File search ignores non-root dirs if prefixed with "*"
+* Treat members of `projectile-globally-ignored-file-suffixes` as file name suffixes (previous treat as file extensions).
 
 ### Bugs fixed